pyproject.toml
[tool.poetry]
name = "safeflat"
version = "0.1.0"
description = ""
authors = ["safeflat"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
langchain-openai = "^0.1.4"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
pyairports = "^2.1.1"
pytest-xdist = "^3.6.1"
pandas-stubs = "^2.2.2.240603"
pycparser-stubs = "^0.0.1"
sortedcontainers-stubs = "^2.4.2"
sqlalchemy-stubs = "^0.4"
tqdm-stubs = "^0.2.1"
instructor = "^1.3.3"

[tool.poetry.group.dev.dependencies]
sphinx = "^5.2.1"
black = "^24.4.2"
pylint = "^3.1.0"
awscli = "^1.32.95"
poetry-plugin-export = "^1.7.1"
pytest = "^8.2.1"
mypy = "^1.10.0"
flake8 = "^7.0.0"
pytest-mock = "^3.14.0"
aws-sam-cli = "^1.119.0"



[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--verbose"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"